Provide a little wchar_t support at the public API boundary to start addressing https://github.com/mono/mono/issues/7117 Chinese characters cause "String conversion error: Illegal byte sequence encounted in the input." #7117
In particular, instead of char* presumed utf8 and easy to get wrong,
some work for users to get correct, provide for wchar_t input, which
is easy to get correct.

Implementation just turns around and translates to utf8, which isn't really ideal
but is a path of significantly less resistance.