Fix race conditions and inefficiencies in mono_arch_get_seq_point_info. Groundwork for more similar.
Specifically:

The race condition is that a failed lookup leads unconditionally
to an insert. The lock is released between the lookup and the insert.
Another insert could occur in between.

The smallest fix is to lookup again before insert, under the second look.
(Another fix is to malloc+insert under first lock, but that would significantly
extend lock hold time, presumably the original author had that in mind).

However the inefficiency is that there are already two lookups (insert implies lookup)
and the fix would add a third. However, this can all be achieved with one lookup.

In general the pattern of lookup + insert can be optimized.
At least for tables that are never deleted from, or only used on single thread (
no intervening delete).
If a table does undergo delete then, one must fallback to the original idea
of an extra lookup.