Steps toward reducing allocations and hash table lookups as requested here:
https://github.com/mono/mono/pull/11941#issuecomment-449157323

This change eliminates the allocations and establishes macros
that can be moved, with some further change, to icall-def.h.

The macros mostly statically initialize the data, and a later
version will do so more, with the notable exception of the "dyn" case.

Once in icall-def.h, the statically allocated MonoJitIcallInfo
can be referenced directly from where hash tables are looked in.