Remove icall hash order sensitivity.
I found this adding some JIT icalls.

Adding JIT icalls causes preexisting ones to fail, not found at AOT load time.

Some icalls are registered with multiple names.

For example:

```
	register_opcode_emulation (OP_FCONV_TO_U8,
		"__emul_fconv_to_u8", "ulong double", mono_fconv_u8, "mono_fconv_u8", FALSE);

	register_opcode_emulation (OP_FCONV_TO_U,
		"__emul_fconv_to_u", "ulong double", mono_fconv_u8, "mono_fconv_u8", TRUE);
```

and `mono_tls_get_lmf_addr` and `mono_get_lmf_addr`.

and then:
```
		/* JIT icall wrappers */
		/* FIXME: locking - this is "safe" as full-AOT threads don't mutate the icall hash*/
		g_hash_table_foreach (mono_get_jit_icall_info (), add_jit_icall_wrapper, acfg);
.
.
.
add_jit_icall_wrapper:
{
	name = g_strdup_printf ("__icall_wrapper_%s", callinfo->name);
	wrapper = mono_marshal_get_icall_wrapper (callinfo->sig, name, callinfo->func, TRUE);
	add_method (acfg, wrapper);
}
.
.
.
mono_marshal_get_icall_wrapper:
...
	if ((res = mono_marshal_find_in_cache (cache, (gpointer) func)))
		return res;
```

The result is that only the first name in hash order of a given function gets a wrapper.
That depends on number of elements in the hash table, which is not the intent.

Add icalls breaks things -- FullAOT is missing wrappers.

Depending on hash order in general is quite bad.

Hash order is not just a function of the names and hash function.
It is also affected by the number of buckets, i.e, the number
of icalls, i.e. rehashing.

Furthermore it does not appear necessary to single instance by function pointer.
It saves approximately two wrappers.

Also bottleneck the duplicated logic.

The multiple instances of should probably also both have or not have wrapper.

Prohibiting functions to be registered twice is also reasonable.