Remove JIT ICall hashing by string and by address, except temporary in asserts. And still, do much less of the lookups and make the inserts cheaper.
This is only JIT icalls. It is not icalls from C#, declared in C#.

Replace uses by direct pointers to data or enums.
The direct pointers could usually be enums if that is preferred.
This change can be split up into smaller pieces.

 - Introduce the enums.
 - Cheapen the inserts. (static allocation)
 - Reduce the lookups. (pointer or enum)
 - Ongoing lookup reduction. (ditto)
 - Until lookups all gone.
 - Remove the inserts.

This is actually mostly very straightforward.
In particular most of the string lookups are trivial to remove.
Most of the call emits are trivial to not lookup.

What is tricker (incorrect?) is the ABS patches.

Pointer vs. enum:
 Pointers are very natural and provide most
   direct access to all the data to the recipient

 Enum is typically smaller (32bits vs. 64 bits)
   and could be even smaller -- 9 bits suffice here.
   And const static arrays of enums do not cost relocation.

 Enum could also be a primary key into multiple arrays,
   if there was not just one data about each icall, as there
   currently is.

 Enums work more naturally for persistance, and are used
   here for AOT. Where previously there were strings.
   Enums make for a more brittle file format however, vs. strings.

 Enums fit better where there are preceding 32bit limits, i.e.
   passing tokens through generated IL. And are used that
   way here. Pointers do work, as previously
   was used here a string and a hash lookup.