[interp] Store pointers to stack in InterpFrame as 32bit offsets from the start of InterpFrame -- InterpFrame is also stack allocated.
As long as stack is smaller than 2GB or 4GB this will work and saves
20 bytes of frame size. (or probably 16 due to alignment).
Or at least the values within 2GB of each other.

This also includes:
 https://github.com/mono/mono/pull/16317
 https://github.com/mono/mono/pull/16316

which remove another 16 bytes.

Unusual architectures with discontiguous stacks will break.

However, this is well abstracted and could be ifdefed/configured.
I don't know how to detect it automatically.