[interp] Do not inline stackval_to_data int MINT_BOX.
With other changes, this has been seen to conserve stack.
Against current master, this does not make a difference.

Likely the changes need to make a difference here are acceptable,
and likely a combined PR will be presented shortly, i.e. to decrease
frame from 0xB8 to 0x98.

An alternative also to be presented shortly is to outline mint_box entirely.